[Color items] Adapts the color item list to the kids/adult flavors.
The ColorItemListWrapper desribes the common contract of the wrapper.
The FlavorColorItemListWrapper defines the specific behaviors.

For the kids flavor, the items are presented in a grid.
Only a preview of the color is presented.

For the adult flavor, the items are presented in a vertical list.
The hexadecimal code id displayed next to the color.
A long press on the row will copy the hexadecimal code to the clipboard.