[Palette list] Adapt the palette list to the kids/adult falvors.
The PaletteListWrapper describes the common contract of the wrapper.
The FlavorPaletteListWrapper defines the specific behaviors.

For the kids flavor, the palettes are presented in a grid.
Only a preview of the palette is presented.

For the adult flavor, the palettes are presented in a vertical list.
The name of the palette is displayed next to the preview.